+++ cmake/modules/AddMPIFlags.cmake
@@ -13,13 +13,6 @@
 # register HAVE_MPI for config.h
 set(HAVE_MPI ${MPI_FOUND})
 
-# register METIS library as dune package
-if(MPI_FOUND)
-  dune_register_package_flags(
-    LIBRARIES MPI::MPI_C
-    COMPILE_DEFINITIONS "ENABLE_MPI=1;MPICH_SKIP_MPICXX;OMPI_SKIP_MPICXX;MPIPP_H;MPI_NO_CPPBIND")
-endif()
-
 # Add function to link targets against MPI library
 function(add_dune_mpi_flags _targets)
   if(MPI_FOUND)