+++ cmake/modules/AddPTScotchFlags.cmake
@@ -13,11 +13,6 @@
 # set HAVE_PTSCOTCH for config.h
 set(HAVE_PTSCOTCH ${PTScotch_FOUND})
 
-# register all PTScotch related flags
-if(PTScotch_FOUND)
-  dune_register_package_flags(LIBRARIES "PTScotch::PTScotch")
-endif()
-
 function(add_dune_ptscotch_flags _targets)
   if(PTScotch_FOUND)
     foreach(_target ${_targets})