+++ cmake/modules/AddTBBFlags.cmake
@@ -10,21 +10,15 @@
 #       A list of targets to use TBB with.
 #
 
-# set HAVE_TBB for config.h
+include_guard(GLOBAL)
 
+# text for feature summary
+set_package_properties("TBB" PROPERTIES
+  DESCRIPTION "Threading Building Blocks library"
+  PURPOSE "Parallel programming on multi-core processors")
+
+# set variable for config.h
 set(HAVE_TBB ${TBB_FOUND})
-
-# perform DUNE-specific setup tasks
-if(TBB_FOUND)
-  set(TBB_CACHE_ALIGNED_ALLOCATOR_ALIGNMENT 128)
-  message(STATUS "defaulting TBB_CACHE_ALIGNED_ALLOCATOR_ALIGNMENT to 128")
-  dune_register_package_flags(
-    COMPILE_DEFINITIONS ENABLE_TBB=1 ${TBB_COMPILE_DEFINITIONS}
-    COMPILE_OPTIONS ${TBB_COMPILE_OPTIONS}
-    INCLUDE_DIRS ${TBB_INCLUDE_DIRS}
-    LIBRARIES ${TBB_LIBRARIES}
-    )
-endif()
 
 # function for adding TBB flags to a list of targets
 function(add_dune_tbb_flags _targets)